Workaround for if Mono 4.2.4 is installed
(Add <clear/> to <webRequestModules> in *.exe.config file; otherwise
Mono complains that the webRequestModule is already registered.)

You don't need Mono installed to use this. We ship MonoKickstart, but
it's a build I made based off of Mono 4.2.4 and unfortunately some
hardcoded paths have crept into the binaries. This seems to be the only
case of conflict.